Reduce application complexity

+++ Intern/rayx-ui/src/RayProcessing.cpp
@@ -111,7 +111,7 @@ void sortRaysByElement(const RAYX::BundleHistory& rays, std::vector<std::vector<
     // Iterate over all rays in the bundle history
     for (const auto& rayBundle : rays) {
         for (const auto& ray : rayBundle) {
-            sortedRays[ray.m_lastElement].push_back(ray);
+            sortedRays[static_cast<size_t>(ray.m_lastElement)].push_back(ray);
         }
     }
 }
